HibernateDemo factory utils

+++ week3/exercises_day_10/HibernateDemo/src/com/revature/App_02.java
@@ -1,6 +1,7 @@
 package com.revature;
 
 import com.revature.beans.Employee;
+import com.revature.util.HibernateUtils;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
@@ -11,9 +12,10 @@ import org.hibernate.boot.registry.StandardServiceRegistryBuilder;
 
 public class App_02 {
     public static void main(String[] args) {
-        StandardServiceRegistry ssr = new StandardServiceRegistryBuilder().configure("hibernate.cfg.xml").build();
-        Metadata metadata = new MetadataSources(ssr).getMetadataBuilder().build();
-        SessionFactory sessionFactory = metadata.getSessionFactoryBuilder().build();
+//        StandardServiceRegistry ssr = new StandardServiceRegistryBuilder().configure("hibernate.cfg.xml").build();
+//        Metadata metadata = new MetadataSources(ssr).getMetadataBuilder().build();
+//        SessionFactory sessionFactory = metadata.getSessionFactoryBuilder().build();
+        SessionFactory sessionFactory = HibernateUtils.getSessionFactory();
         Session session = sessionFactory.openSession();
 
         Transaction transaction = session.beginTransaction();
@@ -27,11 +29,15 @@ public class App_02 {
         session.save(employee3);
 
         Employee employee = session.find(Employee.class, 2);
-        employee.setFirstName("My");
-        employee.setLastName("Mom");
-        employee.setFirstName("Your");
-        session.delete(employee);
+        if(employee != null) {
+            employee.setFirstName("My");
+            employee.setLastName("Mom");
+            employee.setFirstName("Your");
+            session.delete(employee);
+        }
 
         transaction.commit();
+
+        session.close();
     }
 }
